Implemented check evasions in movegen ( nodes 187971 nps 1068017 os darwin )

+++ src/core/perft.cpp
@@ -5,12 +5,12 @@
 
 namespace sagittar::perft {
 
-    u64 perft(const Position& pos, const Depth depth) {
+    size_t perft(const Position& pos, const Depth depth) {
         if (depth == 0)
         {
             return 1ULL;
         }
-        u64                         nodes = 0ULL;
+        size_t                      nodes = 0ULL;
         containers::ArrayList<Move> moves;
         pseudolegalMoves<MovegenType::ALL>(&moves, pos);
         for (auto const& move : moves)
@@ -24,13 +24,13 @@ namespace sagittar::perft {
         return nodes;
     }
 
-    u64 divide(const Position& pos, const Depth depth) {
+    size_t divide(const Position& pos, const Depth depth) {
         if (depth == 0)
         {
             return 1ULL;
         }
-        u64                         nodes       = 0ULL;
-        u64                         total_nodes = 0ULL;
+        size_t                      nodes       = 0ULL;
+        size_t                      total_nodes = 0ULL;
         containers::ArrayList<Move> moves;
         pseudolegalMoves<MovegenType::ALL>(&moves, pos);
         for (auto const& move : moves)
@@ -41,10 +41,10 @@ namespace sagittar::perft {
                 nodes = perft(pos_copy, depth - 1);
                 total_nodes += nodes;
                 move.display();
-                std::cout << " " << (u64) nodes << std::endl;
+                std::cout << " " << (size_t) nodes << std::endl;
             }
         }
-        std::cout << std::endl << "Perft = " << (u64) total_nodes << std::endl;
+        std::cout << std::endl << "Perft = " << (size_t) total_nodes << std::endl;
         return total_nodes;
     }
